@@ -1,10 +1,10 @@
 {
-  "name": "Raptors",
+  "name": "saller",
   "main": "server.js",
-  "repository": "git@github.com:Libbay/Raptors.git",
+  "repository": "git@github.com:sallerjs/saller.git",
   "scripts": {
     "start": "better-npm-run start-dev",
-    "start:api": "concurrently --kill-others \"better-npm-run start-dev\" \"better-npm-run start-dev-api\"",
+    "api": "better-npm-run start-dev-api",
     "build:test": "better-npm-run build:test",
     "build:prod": "better-npm-run build:prod"
   },
@@ -63,7 +63,7 @@
     "autoprefixer": "^6.1.2",
     "autoprefixer-loader": "~2.0.0",
     "babel-cli": "^6.11.4",
-    "babel-eslint": "^3.1.23",
+    "babel-eslint": "^6.1.2",
     "babel-loader": "^6.2.3",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.9.0",
@@ -77,9 +77,11 @@
     "compression-webpack-plugin": "^0.2.0",
     "concurrently": "^2.2.0",
     "css-loader": "^0.14.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^3.1.0",
-    "eslint-plugin-react": "^3.15.0",
+    "eslint": "^3.3.1",
+    "eslint-config-airbnb": "^10.0.1",
+    "eslint-plugin-import": "^1.13.0",
+    "eslint-plugin-jsx-a11y": "^2.1.0",
+    "eslint-plugin-react": "^6.1.1",
     "express": "^4.13.3",
     "extract-text-webpack-plugin": "^0.8.1",
     "file-loader": "^0.8.5",
@@ -90,6 +92,8 @@
     "jsx-loader": "^0.13.2",
     "less": "~2.5.1",
     "less-loader": "^2.2.0",
+    "mongoose": "~3.6.13",
+    "morgan": "~1.0.0",
     "node-libs-browser": "^0.5.2",
     "open": "0.0.5",
     "postcss-loader": "^0.8.0",
@@ -98,10 +102,12 @@
     "redux-logger": "^2.1.4",
     "redux-simple-router": "^0.0.10",
     "redux-thunk": "^1.0.0",
+    "standard": "^7.1.2",
     "stats-webpack-plugin": "^0.2.2",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.7",
     "webpack": "^1.9.10",
+    "webpack-dashboard": "0.0.1-alpha.1",
     "webpack-dev-middleware": "^1.2.0",
     "webpack-dev-server": "^1.9.0",
     "webpack-hot-middleware": "^2.2.0"
