@@ -1,11 +1,2 @@
-#!/usr/bin/env node
-if (process.env.NODE_ENV !== 'production') {
-  if (!require('piping')({
-    hook: true,
-    ignore: /(\/\.|~$|\.json$)/i
-  })) {
-    return;
-  }
-}
-require('../server.babel'); // babel registration (runtime transpilation for node)
-require('../api/api');
+// babel registration (runtime transpilation for node)
+require('../api/api')
